Changes before error encountered
Agent-Logs-Url: https://github.com/understyx/Vk-edgy-wotlk/sessions/b3ac3800-8933-45ec-89ba-de1001cec1b7

+++ src/wowmemory/wowmemory.cpp
@@ -171,7 +171,7 @@ static uint32_t ReadAurasForUnit(uintptr_t unitBase, AuraInfo* outAuras, size_t 
             uintptr_t stackAddr = entryBase + WoWOffsets::auraStructStackCountOffset;
             if (IsReadableRange(stackAddr, sizeof(uint8_t))) {
                 uint8_t sc = *reinterpret_cast<const uint8_t*>(stackAddr);
-                info.stackCount = sc > 0 ? sc : 1u;
+                info.stackCount = sc > 0 ? sc : static_cast<uint8_t>(1);
             } else {
                 info.stackCount = 1;
             }
